requirements.txt
Pillow
screeninfo